Extract post details compoent
